debug

+++ paperstream/feeds/tasks.py
@@ -23,7 +23,8 @@ def reset_stream(user_pk, stream_name='main', restrict_journal=False):
         feed.add_papers_seed(user.lib.papers.all())
     # reset
     feed.clear_all()
-    user.streamlayout.delete()
+    if user.streamlayout:
+        user.streamlayout.delete()
     # update
     feed.update(restrict_journal=restrict_journal)
     return user_pk
@@ -55,7 +56,8 @@ def reset_trend(user_pk, trend_name='main'):
     df, _ = Trend.objects.get_or_create(user_id=user_pk, name=trend_name)
     user = User.objects.get(pk=user_pk)
     # reset filter
-    user.trendlayout.delete()
+    if user.trendlayout:
+        user.trendlayout.delete()
     # update
     df.update()
     return user_pk
